Extracted remote code fetching logic to hook
