+++ src/editor/controls/GitHubRepositoryControl.tsx
@@ -1,8 +1,8 @@
 import {BaseControl, Button, TextControl,} from '@wordpress/components';
 import {useRef} from '@wordpress/element';
 import {__} from '@wordpress/i18n';
-import apiFetch from '@wordpress/api-fetch';
-import useSWR, { mutate } from "swr";
+import {useRemoteCodeFetching} from "../../hooks/useRemoteCodeFetching";
+import {extractLineNumbers, isValidUrl} from "../../util/urls";
 
 interface GitHubRepositoryControlProps {
     value: string;
@@ -10,18 +10,9 @@ interface GitHubRepositoryControlProps {
     onCodeFetched: (data: any) => void;
 }
 
-const fetcher = async (url: string) => {
-    const response = await apiFetch({
-        path: `code-block-pro/v1/code?url=${encodeURIComponent(url)}`,
-        method: 'GET',
-    });
-
-    return response.code;
-};
-
 export const GitHubRepositoryControl = ({value, onChange, onCodeFetched}: GitHubRepositoryControlProps) => {
     const inputRef = useRef<HTMLInputElement>(null);
-    const { data, error } = useSWR(isValidUrl(value) ? value : null, fetcher);
+    const { data, error, mutate } = useRemoteCodeFetching(value);
 
     if (data && !error) {
         if (value.indexOf("#L")) {
@@ -65,53 +56,3 @@ export const GitHubRepositoryControl = ({value, onChange, onCodeFetched}: GitHub
         </BaseControl>
     );
 };
-
-function isValidUrl(str: string) {
-    try {
-        const url = new URL(str);
-
-        return url.host === 'raw.githubusercontent.com' ||
-            url.host === 'gist.githubusercontent.com' ||
-            url.host === 'github.com';
-    } catch (e) {
-        return false;
-    }
-}
-
-function extractLineNumbers(url) {
-    const hashIndex = url.indexOf("#L");
-    if (hashIndex === -1) {
-        return null;
-    }
-
-    const strAfterHash = url.substring(hashIndex + 1);
-    const strSplitByDash = strAfterHash.split('-');
-    const lineNumbers = strSplitByDash.map(getLineNumberFromToken);
-
-    const isFirstNumberValid = lineNumbers[0] > 0;
-    const isSecondNumberValid = lineNumbers[1] > 0;
-    if (!isFirstNumberValid) {
-        return null;
-    }
-
-    return {
-        startLine: lineNumbers[0],
-        endLine: isSecondNumberValid ? lineNumbers[1] : lineNumbers[0],
-    };
-}
-
-function getLineNumberFromToken(token: string) {
-    const splitByColumn = token.split('C');
-    const lineNumber = splitByColumn[0].substring(1);
-
-    return parseInt(lineNumber);
-}
-
-async function fetchFile(url: string) {
-    const response = await apiFetch({
-        path: `code-block-pro/v1/code?url=${encodeURIComponent(url)}`,
-        method: 'GET',
-    });
-
-    return response.code;
-}
